TDD PIG LATIN
Task: Use TDD practices to build a function that translates from English to Pig Latin.

Build Specifications
Translation must all be done in a translate() method that takes a string parameter and returns the translated Pig Latin string.
The translation functionality must be developed with TDD using tests on the translate() method.
If desired, you may also create other helper functions and optionally test them.

Additional Build Specifications
Convert each word to lowercase before translating.
If a word starts with a vowel, just add “way” onto the ending.
If a word starts with a consonant, move all of the consonants that appear before the first vowel to the end of the word, then add “ay” to the end of the word.